Move DecouplingFieldInfosCodec to upper level
